+++ cpp-module-09/ex01/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: dnikifor <dnikifor@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/22 14:52:56 by dnikifor          #+#    #+#             */
-/*   Updated: 2024/08/22 17:15:31 by dnikifor         ###   ########.fr       */
+/*   Updated: 2024/08/22 20:27:38 by dnikifor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@
 
 class RPN {
 	private:
-		static std::stack<int>	_stack;
-		static bool				isOperation(char c);
-		static int				executeOperation(int a, int b, char c);
+		static std::stack<float>	_stack;
+		static bool					isOperation(char c);
+		static float				executeOperation(float a, float b, char c);
 
 		RPN() = delete;
 		RPN(const RPN &) = delete;
 		RPN& operator=(const RPN &) = delete;
 		~RPN() = delete;
 	public:
-		static void				execute(char *arg);
+		static void					execute(char *arg);
 };
